// src/shared/api/models/postApiCampaignsStatistics200Item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type PostApiCampaignsStatistics200Item = {
    /** Number of messages sent */
    messageSent: number;
    /** Number of messages not delivered */
    notDelivered: number;
    /** Email open rate percentage */
    openRate: string;
    /** Total email click count */
    clicks: number;
    /** Revenue generated */
    revenue: string;
    /** Number of unsubscribed recipients */
    unsubscribed: number;
};